Beads unsupervised classification

This function uses clustering to classify the events into beads or cells. Beads are expected to be a group of very little dispersion, so the classification is assessed by looking into the Coefficient of Variation of the rows classified as beads. If this is > CV_max, the classification failed and the df is converted to 0. Otherwise, df with a new class column is outputted. The summary stats are written in CSV files either case.

BeadsUnsup(
  df,
  method = c("GMM", "k_means"),
  beads_ch = beads_ch,
  n_clusters = 2,
  CV_max = 0.05,
  filename = filename,
  output_path = output_path,
  FCSdesc = FCSdesc
)

Arguments

df
An object of class data.frame.

method
A character vector with the name of the clustering algorithm to be used. Options are Gaussian Mixture Models 'GMM' or K-means 'k_means'.

beads_ch
A numeric vector with the column indexes for the beads channels.

n_clusters
A numeric vector indicating the number of clusters to be targeted.

CV_max
A numeric vector indicating the maximum value for the coefficient of variation (CV) to evaluate beads cluster.

filename
A character vector with the name of the file that originated the data.frame.

output_path
A character vector with full path names.

FCSdesc
FCS column descriptions (extracted from flowFrame object)

Value

If CV< CV_max returns a data.frame with new class column for beads or cells. Otherwise, returns a vector = 0.